Make map sources cacheable, so that the front page loads a bit faster.
Also move map-sources tests from the images app to maps.
