+++ project/images/tests/test_source_other.py
@@ -194,155 +194,6 @@ class SourceListTest(ClientTest):
         )
 
 
-@override_settings(MAP_IMAGE_COUNT_TIERS=[2, 3, 5])
-class SourceMapTest(ClientTest):
-    """
-    Test the map on the source list page.
-    """
-    @classmethod
-    def setUpTestData(cls):
-        super().setUpTestData()
-
-        cls.user = cls.create_user()
-
-    def test_map_sources_all_fields(self):
-        source = self.create_source(
-            self.user, visibility=Source.VisibilityTypes.PUBLIC)
-        for _ in range(2):
-            self.upload_image(self.user, source)
-
-        self.client.force_login(self.user)
-        response = self.client.get(reverse('source_list'))
-
-        self.assertDictEqual(
-            response.context['map_sources'][0],
-            dict(
-                sourceId=source.id,
-                latitude=source.latitude,
-                longitude=source.longitude,
-                type='public',
-                size=1,
-                detailBoxUrl=reverse('source_detail_box', args=[source.pk]),
-            ),
-        )
-
-    def test_map_sources_type_field(self):
-        # Test both possible type values.
-        public_source = self.create_source(
-            self.user, visibility=Source.VisibilityTypes.PUBLIC)
-        for _ in range(2):
-            self.upload_image(self.user, public_source)
-
-        private_source = self.create_source(
-            self.user, visibility=Source.VisibilityTypes.PRIVATE)
-        for _ in range(2):
-            self.upload_image(self.user, private_source)
-
-        self.client.force_login(self.user)
-        response = self.client.get(reverse('source_list'))
-
-        # Set comparison, since source order is not defined here.
-        ids_and_types = {
-            (d['sourceId'], d['type'])
-            for d in response.context['map_sources']
-        }
-        self.assertSetEqual(
-            ids_and_types,
-            {
-                (public_source.pk, 'public'),
-                (private_source.pk, 'private'),
-            },
-        )
-
-    def test_map_sources_size_field(self):
-        """Test all possible source size tiers (size meaning image count)."""
-        source_tier_0 = self.create_source(self.user)
-        self.upload_image(self.user, source_tier_0)
-
-        source_tier_1 = self.create_source(self.user)
-        for _ in range(2):
-            self.upload_image(self.user, source_tier_1)
-
-        source_tier_2 = self.create_source(self.user)
-        for _ in range(4):
-            self.upload_image(self.user, source_tier_2)
-
-        source_tier_3 = self.create_source(self.user)
-        for _ in range(5):
-            self.upload_image(self.user, source_tier_3)
-
-        self.client.force_login(self.user)
-        response = self.client.get(reverse('source_list'))
-
-        # Set comparison, since source order is not defined here.
-        # Note that source_tier_0 is not in the map sources.
-        ids_and_sizes = {
-            (d['sourceId'], d['size'])
-            for d in response.context['map_sources']
-        }
-        self.assertSetEqual(
-            ids_and_sizes,
-            {
-                (source_tier_1.pk, 1),
-                (source_tier_2.pk, 2),
-                (source_tier_3.pk, 3),
-            },
-        )
-
-    def test_exclude_test_sources(self):
-        source_1 = self.create_source(self.user, name="Source 1")
-        for _ in range(2):
-            self.upload_image(self.user, source_1)
-
-        test_source_1 = self.create_source(self.user, name="Test 1")
-        for _ in range(2):
-            self.upload_image(self.user, test_source_1)
-
-        test_source_2 = self.create_source(
-            self.user, name="User's temporary source")
-        for _ in range(2):
-            self.upload_image(self.user, test_source_2)
-
-        self.client.force_login(self.user)
-        response = self.client.get(reverse('source_list'))
-
-        ids = {d['sourceId'] for d in response.context['map_sources']}
-        self.assertSetEqual(ids, {source_1.pk})
-
-    def test_exclude_no_lat_or_long(self):
-        # One 0 is OK
-        source_1 = self.create_source(self.user, latitude='0')
-        for _ in range(2):
-            self.upload_image(self.user, source_1)
-
-        # Either blank: exclude.
-        # This is only possible for old sources that were created before
-        # lat/long became required in the form.
-        test_source_1 = self.create_source(self.user)
-        test_source_1.latitude = ''
-        test_source_1.save()
-        for _ in range(2):
-            self.upload_image(self.user, test_source_1)
-
-        test_source_2 = self.create_source(self.user)
-        test_source_2.longitude = ''
-        test_source_2.save()
-        for _ in range(2):
-            self.upload_image(self.user, test_source_2)
-
-        # Both 0s: exclude
-        test_source_3 = self.create_source(
-            self.user, latitude='0', longitude='0')
-        for _ in range(2):
-            self.upload_image(self.user, test_source_3)
-
-        self.client.force_login(self.user)
-        response = self.client.get(reverse('source_list'))
-
-        ids = {d['sourceId'] for d in response.context['map_sources']}
-        self.assertSetEqual(ids, {source_1.pk})
-
-
 class SourceDetailBoxTest(ClientTest):
     """
     Test the map's source detail popup box.
