+++ project/images/views.py
@@ -32,7 +32,7 @@ from lib.decorators import (
     source_permission_required,
     source_visibility_required,
 )
-from map.utils import get_map_sources
+from map.utils import cacheable_map_sources
 from newsfeed.models import NewsItem
 from vision_backend.models import Classifier
 from vision_backend.utils import reset_features
@@ -73,7 +73,7 @@ def source_list(request):
 
     return render(request, 'images/source_list.html', {
         'your_sources': your_sources_dicts,
-        'map_sources': get_map_sources(),
+        'map_sources': cacheable_map_sources.get(),
         'other_public_sources': other_public_sources,
         'total_sources': total_sources,
         'total_images': total_images,
